requirements.txt
torch
torchvision
ultralytics
opencv-python
matplotlib
numpy
pillow
tqdm
